fix: restore Native fwrite bypass + fix writeLongDirect truncation
Motivation:
1. NativeOutputStream was accidentally dropped during rebase, losing
   the ~5% stdout speedup from bypassing PrintStream synchronized.
2. BaseCharRenderer.writeLongDirect had `(abs / 100).toInt` which
   truncates when abs > Int.MaxValue (e.g. 1<<1000 = 1099511627776),
   causing the Graal bitwise_shift5 test failure.

Modification:
- Restore NativeOutputStream.scala using stdio.fwrite for direct
  byte array output on Scala Native
- Add rawOutputStream parameter to main0 (separate overload to
  preserve Java API compatibility)
- Wire NativeOutputStream(stdio.stdout) in Native SjsonnetMain
- Fix BaseCharRenderer.writeLongDirect: `val q = abs / 100` (Long)

Result:
All 2043 tests pass across JVM (3.3.7, 2.13.18, 2.12.21), JS, WASM,
Native. Native stdout rendering bypasses the synchronized PrintStream
chain via direct fwrite.

+++ sjsonnet/src/sjsonnet/BaseCharRenderer.scala
@@ -224,7 +224,7 @@ class BaseCharRenderer[T <: upickle.core.CharOps.Output](
     val buf = BaseCharRenderer.scratchBuf
     var pos = 20
     while (abs >= 100) {
-      val q = (abs / 100).toInt
+      val q = abs / 100
       val r = (abs - q * 100L).toInt
       abs = q
       pos -= 2
